Plotting and Exploration Experiment - Added MatLab files to plot fixed and exploration data. -> To be refactored - Added experiment script to iterate through exploration rates. - Modified fixed experiment script to run through trials to recreate AAMAS results.

+++ MultiNightBar.hpp
@@ -220,7 +220,7 @@ void printVector(std::vector<T> v){
 
     std::cout << v[0];
 
-    for (int i = 1; i < v.size(); ++i){
+    for (size_t i = 1; i < v.size(); ++i){
         std::cout << ", " << v[i];
     }
     std::cout << "\n";
